+++ desktop/packages/windows-utils/windows-utils-rs/lib.rs
@@ -1,0 +1,12 @@
+#![cfg(target_os = "windows")]
+
+use neon::{prelude::ModuleContext, result::NeonResult};
+
+mod shortcut;
+
+#[neon::main]
+fn main(mut cx: ModuleContext<'_>) -> NeonResult<()> {
+    cx.export_function("readShortcut", shortcut::read_shortcut)?;
+
+    Ok(())
+}